Add -i, --inverse option
Add -i, --inverse option to invert the meaning of bare (no "=") lump
names passed to a list of changes. Normally all lumps are assumed and
bare lumps are removed. With -i, --inverse there are no lumps by
default, but bare lumps are allowed. Add unit test w-in-dir-changes-inv
to test it.